fix(web): review round 3 — org-switch refetch, legacy redirect, Node pin, etc.
- P1 org switch: switchOrg now `clear()` THEN `await refetch()` — clear() does
  not refetch the active ['me'] observer, so the shell could keep showing the
  previous org. (auth-provider)
- P1 legacy deep links: /sessions/:sandboxId redirects to /sandboxes/:sandboxId
  (was falling through to "/"). (App.tsx)
- P1/P2 Node/Vite 8: Vite 8 needs Node ≥20.19; pin engines + web/.nvmrc(22) +
  deploy-server workflow to Node 22 (deploy-api-edge already 22).
- P2 Terminal: guard the WebSocket onopen/onmessage/onclose/onerror with
  `disposed` so a late close can't setState/write to a disposed xterm.
- P3 LogsPanel: fix the stale "server-side filtering" contract comment (it's
  client-side now).
- P3: move the ~300 lines of dead agent API out of client.ts into a dormant
  src/api/agents.ts (nothing imports it; client.ts 651 → 340 lines).
- docs: note the core-lib upgrade was intentional + the Node pin.
